Teleporters no longer teleport multi-way blocks.
Fixed a minor error in disabling test/load/clear after opening the load level dialog and canceling (it wasn't considering if the already-loaded level was valid).

Refactored the word Song back to sound in many files. Removed the broken crate object and image, these will later have animations that don't stay on the map (it was also possible to be teleported and moved by a belt on the teleporter in a single frame, rare & harmless incongruency in behavior).

The active item indicator in level editor is now cleared when loading/clearing levels.

Teleporters, filters (and now rotaters) being pasted in the level editor now follow expected logic. Pasting/clicking to set them on top of each other only works on initial paste/click, so dragging will not make overlapping tiles for these arbitrarily stackable types (to prevent accidental pasting).